Mostly done clean structure
Still lacking some features from the propertygroup.
Also casts runtime error "Error: SQLite Error 1: 'no such table: Authors'."
