+++ src/Chirp.Core/Cheep.cs
@@ -1,0 +1,8 @@
+namespace Chirp.Core;
+public class Cheep
+{
+    public int CheepId { get; set; }
+    public required Author Author { get; set; }
+    public required string Text { get; set; }
+    public DateTime TimeStamp { get; set; }
+}
