+++ src/Chirp.Core/CheepViewModel.cs
@@ -1,0 +1,2 @@
+namespace Chirp.Core; 
+public record CheepViewModel(string Author, string Message, string TimeStamp);
