+++ src/Chirp.Core/ICheepService.cs
@@ -1,0 +1,6 @@
+namespace Chirp.Core;
+public interface ICheepService
+{
+    public List<CheepViewModel> GetCheeps();
+    public List<CheepViewModel> GetCheepsFromAuthor(string authorName);
+}
