+++ src/Chirp.Infrastructure/ChirpDBContext.cs
@@ -1,6 +1,7 @@
+using Chirp.Core;
 using Microsoft.EntityFrameworkCore;
 
-namespace Chirp.Razor;
+namespace Chirp.Infrastructure;
 
 public class ChirpDBContext : DbContext
 {
@@ -21,21 +22,4 @@ public class ChirpDBContext : DbContext
     // special "local" folder for your platform.
     protected override void OnConfiguring(DbContextOptionsBuilder options)
         => options.UseSqlite($"Data Source={DbPath}");
-}
-
-public class Author
-{
-    public int AuthorId { get; set; }
-    public required string Name { get; set; }
-    public required string Email { get; set; }
-
-    public List<Cheep> Cheeps { get; set; } = new();
-}
-
-public class Cheep
-{
-    public int CheepId { get; set; }
-    public required Author Author { get; set; }
-    public required string Text { get; set; }
-    public DateTime TimeStamp { get; set; }
 }
