+++ src/Chirp.Web/Pages/Public.cshtml
@@ -1,6 +1,7 @@
 ﻿@page "/"
-@using Chirp.Razor.Pages
-@model Chirp.Razor.Pages.PublicModel
+@using Chirp.Core
+@using Chirp.Web.Pages
+@model Chirp.Web.Pages.PublicModel
 @{
     ViewData["Title"] = "Chirp!";
     Layout = "Shared/_Layout";
