+++ src/Chirp.Web/Pages/UserTimeline.cshtml
@@ -1,6 +1,7 @@
 ﻿@page "/{author}"
-@using Chirp.Razor.Pages
-@model Chirp.Razor.Pages.UserTimelineModel
+@using Chirp.Core
+@using Chirp.Web.Pages
+@model Chirp.Web.Pages.UserTimelineModel
 @{
     ViewData["Title"] = "Chirp!";
     Layout = "Shared/_Layout";
