+++ src/Chirp.Web/Program.cs
@@ -1,2 +1,36 @@
-﻿// See https://aka.ms/new-console-template for more information
-Console.WriteLine("Hello, World!");
+using Chirp.Core;
+using Chirp.Infrastructure;
+using Microsoft.EntityFrameworkCore;
+
+using (ChirpDBContext db = new ChirpDBContext())
+{
+    // create a database if none exists
+    db.Database.Migrate();
+    // fill the database with data if it's empty
+    DbInitializer.SeedDatabase(db);
+}
+
+WebApplicationBuilder builder = WebApplication.CreateBuilder(args);
+
+// Add services to the container.
+builder.Services.AddRazorPages();
+builder.Services.AddSingleton<ICheepService, CheepService>();
+
+WebApplication app = builder.Build();
+
+// Configure the HTTP request pipeline.
+if (!app.Environment.IsDevelopment())
+{
+    app.UseExceptionHandler("/Error");
+    // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
+    app.UseHsts();
+}
+
+app.UseHttpsRedirection();
+app.UseStaticFiles();
+
+app.UseRouting();
+
+app.MapRazorPages();
+
+app.Run();
